update to community.
Sorry Tom, I've been working outside of GitHub and Transferring my files
in so this update may undo what you did to the folder. But I did check
out the changes you made and it does make the package easier to view so
I'll toss them back into a folder like you did. I'll replace the
software engineer object into the program class too. I'll start working
on the files from GitHub directly so I don't undo anything in the
future!

Updates:
a bunch of variables in the Character class.
experience and level variables.
a move variable from Tyler's gameBoard.
added images in the form of string. When the classes are accessed in
WPF, you can use the variable as the source for the picture location.
getters and setters for all the new variables.
updated toString.

On the hero class, I switched "String Gender" to "Boolean Male".
updated the toString() on that class.

On the program class, I added a hero object that grabs info from the
character class. will do the same for the job role classes.