update to community
brought in the rest of the variables and methods from Tyler's character
class.
added more content for hero, enemy, and software engineer.

will be adding content to the rest of the jobs in the same manner as the

software engineer. Then adding battle, level, experience methods for the
board-game levels.