Game linked together working
File paths were updated in the code to work for the game board and some
of the world map pictures. It now runs again. I've also reorganized the
files of the pictures so they are where they need to be to work, and
removed all the unneeded duplicates that we put everywhere trying to
make it work.